+++ src/dj_settings/__init__.py
@@ -1,3 +1,3 @@
-from dj_settings.utils import SettingsParser, setting
+from dj_settings.settings import SettingsParser, setting
 
 __all__ = ["SettingsParser", "setting"]